refactored torchvision_image_classification. Had to include a subclass of ModelMetadata for this test file to account for extra attribute of weight_names. Also set default required_pcc to 0.96 in subclass. Changed skipping test logic in test functino for vit_h_14 into skip_full_eval_test(). Editted file name too
