+++ tests/models/torchvision/test_torchvision_image_classification.py
@@ -6,21 +6,29 @@ from PIL import Image
 import torch
 import requests
 import pytest
-from tests.utils import ModelTester
-from tt_torch.tools.utils import CompilerConfig, CompileDepth, OpByOpBackend
+from tests.utils import ModelTester, skip_full_eval_test
+from tt_torch.tools.utils import (
+    CompilerConfig,
+    CompileDepth,
+    OpByOpBackend,
+    ModelMetadata,
+)
+
+
+# TorchVision (TV) model metadata
+# Special Subclass of ModelMetada made for this test to include weight_names attr
+class TVModelMetadata(ModelMetadata):
+    def __init__(self, model_name, weights_name, required_pcc=0.96, **kwargs):
+        super().__init__(model_name=model_name, required_pcc=required_pcc, **kwargs)
+        self.weights_name = weights_name
 
 
 class ThisTester(ModelTester):
-    # pass model_info instead of model_name
-    def __init__(self, model_info, mode, *args, **kwargs):
-        # model name in model_info[0]
-        self.model_info = model_info
-        super().__init__(model_info[0], mode, *args, **kwargs)
-
     def _load_model(self):
-        model_name, weights_name = self.model_info
-        self.weights = getattr(models, weights_name).DEFAULT
-        model = models.get_model(model_name, weights=self.weights).to(torch.bfloat16)
+        self.weights = getattr(models, self.model_info.weights_name).DEFAULT
+        model = models.get_model(self.model_name, weights=self.weights).to(
+            torch.bfloat16
+        )
         return model
 
     def _load_inputs(self):
@@ -33,127 +41,142 @@ class ThisTester(ModelTester):
         return batch_t
 
 
-# List of model information tuples (model_name, weights_name)
-model_info_list = [
-    ("googlenet", "GoogLeNet_Weights"),
-    ("densenet121", "DenseNet121_Weights"),
-    ("densenet161", "DenseNet161_Weights"),
-    ("densenet169", "DenseNet169_Weights"),
-    ("densenet201", "DenseNet201_Weights"),
-    ("mobilenet_v2", "MobileNet_V2_Weights"),
-    ("mobilenet_v3_small", "MobileNet_V3_Small_Weights"),
-    ("mobilenet_v3_large", "MobileNet_V3_Large_Weights"),
-    ("resnet18", "ResNet18_Weights"),
-    ("resnet34", "ResNet34_Weights"),
-    ("resnet50", "ResNet50_Weights"),
-    ("resnet101", "ResNet101_Weights"),
-    ("resnet152", "ResNet152_Weights"),
-    ("resnext50_32x4d", "ResNeXt50_32X4D_Weights"),
-    ("resnext101_32x8d", "ResNeXt101_32X8D_Weights"),
-    ("resnext101_64x4d", "ResNeXt101_64X4D_Weights"),
-    ("vgg11", "VGG11_Weights"),
-    ("vgg11_bn", "VGG11_BN_Weights"),
-    ("vgg13", "VGG13_Weights"),
-    ("vgg13_bn", "VGG13_BN_Weights"),
-    ("vgg16", "VGG16_Weights"),
-    ("vgg16_bn", "VGG16_BN_Weights"),
-    ("vgg19", "VGG19_Weights"),
-    ("vgg19_bn", "VGG19_BN_Weights"),
-    ("vit_b_16", "ViT_B_16_Weights"),
-    ("vit_b_32", "ViT_B_32_Weights"),
-    ("vit_l_16", "ViT_L_16_Weights"),
-    ("vit_l_32", "ViT_L_32_Weights"),
-    ("vit_h_14", "ViT_H_14_Weights"),
-    ("wide_resnet50_2", "Wide_ResNet50_2_Weights"),
-    ("wide_resnet101_2", "Wide_ResNet101_2_Weights"),
-    ("regnet_y_400mf", "RegNet_Y_400MF_Weights"),
-    ("regnet_y_800mf", "RegNet_Y_800MF_Weights"),
-    ("regnet_y_1_6gf", "RegNet_Y_1_6GF_Weights"),
-    ("regnet_y_3_2gf", "RegNet_Y_3_2GF_Weights"),
-    ("regnet_y_8gf", "RegNet_Y_8GF_Weights"),
-    ("regnet_y_16gf", "RegNet_Y_16GF_Weights"),
-    ("regnet_y_32gf", "RegNet_Y_32GF_Weights"),
-    ("regnet_y_128gf", "RegNet_Y_128GF_Weights"),
-    ("regnet_x_400mf", "RegNet_X_400MF_Weights"),
-    ("regnet_x_800mf", "RegNet_X_800MF_Weights"),
-    ("regnet_x_1_6gf", "RegNet_X_1_6GF_Weights"),
-    ("regnet_x_3_2gf", "RegNet_X_3_2GF_Weights"),
-    ("regnet_x_8gf", "RegNet_X_8GF_Weights"),
-    ("regnet_x_16gf", "RegNet_X_16GF_Weights"),
-    ("regnet_x_32gf", "RegNet_X_32GF_Weights"),
-    ("swin_t", "Swin_T_Weights"),
-    ("swin_s", "Swin_S_Weights"),
-    ("swin_b", "Swin_B_Weights"),
-    ("swin_v2_t", "Swin_V2_T_Weights"),
-    ("swin_v2_s", "Swin_V2_S_Weights"),
-    ("swin_v2_b", "Swin_V2_B_Weights"),
+# List of Torch Vision Model Metadatas (TVModelMetadata)
+TORCHVISION_MODELS = [
+    TVModelMetadata(model_name="googlenet", weights_name="GoogLeNet_Weights"),
+    TVModelMetadata(model_name="densenet121", weights_name="DenseNet121_Weights"),
+    TVModelMetadata(model_name="densenet161", weights_name="DenseNet161_Weights"),
+    TVModelMetadata(model_name="densenet169", weights_name="DenseNet169_Weights"),
+    TVModelMetadata(model_name="densenet201", weights_name="DenseNet201_Weights"),
+    TVModelMetadata(model_name="mobilenet_v2", weights_name="MobileNet_V2_Weights"),
+    TVModelMetadata(
+        model_name="mobilenet_v3_small", weights_name="MobileNet_V3_Small_Weights"
+    ),
+    TVModelMetadata(
+        model_name="mobilenet_v3_large", weights_name="MobileNet_V3_Large_Weights"
+    ),
+    TVModelMetadata(model_name="resnet18", weights_name="ResNet18_Weights"),
+    TVModelMetadata(model_name="resnet34", weights_name="ResNet34_Weights"),
+    TVModelMetadata(model_name="resnet50", weights_name="ResNet50_Weights"),
+    TVModelMetadata(model_name="resnet101", weights_name="ResNet101_Weights"),
+    TVModelMetadata(model_name="resnet152", weights_name="ResNet152_Weights"),
+    TVModelMetadata(
+        model_name="resnext50_32x4d", weights_name="ResNeXt50_32X4D_Weights"
+    ),
+    TVModelMetadata(
+        model_name="resnext101_32x8d", weights_name="ResNeXt101_32X8D_Weights"
+    ),
+    TVModelMetadata(
+        model_name="resnext101_64x4d", weights_name="ResNeXt101_64X4D_Weights"
+    ),
+    TVModelMetadata(model_name="vgg11", weights_name="VGG11_Weights"),
+    TVModelMetadata(model_name="vgg11_bn", weights_name="VGG11_BN_Weights"),
+    TVModelMetadata(model_name="vgg13", weights_name="VGG13_Weights"),
+    TVModelMetadata(model_name="vgg13_bn", weights_name="VGG13_BN_Weights"),
+    TVModelMetadata(model_name="vgg16", weights_name="VGG16_Weights"),
+    TVModelMetadata(model_name="vgg16_bn", weights_name="VGG16_BN_Weights"),
+    TVModelMetadata(model_name="vgg19", weights_name="VGG19_Weights"),
+    TVModelMetadata(model_name="vgg19_bn", weights_name="VGG19_BN_Weights"),
+    TVModelMetadata(model_name="vit_b_16", weights_name="ViT_B_16_Weights"),
+    TVModelMetadata(model_name="vit_b_32", weights_name="ViT_B_32_Weights"),
+    TVModelMetadata(model_name="vit_l_16", weights_name="ViT_L_16_Weights"),
+    TVModelMetadata(model_name="vit_l_32", weights_name="ViT_L_32_Weights"),
+    TVModelMetadata(
+        model_name="vit_h_14", weights_name="ViT_H_14_Weights", assert_pcc=False
+    ),
+    TVModelMetadata(
+        model_name="wide_resnet50_2", weights_name="Wide_ResNet50_2_Weights"
+    ),
+    TVModelMetadata(
+        model_name="wide_resnet101_2", weights_name="Wide_ResNet101_2_Weights"
+    ),
+    TVModelMetadata(model_name="regnet_y_400mf", weights_name="RegNet_Y_400MF_Weights"),
+    TVModelMetadata(model_name="regnet_y_800mf", weights_name="RegNet_Y_800MF_Weights"),
+    TVModelMetadata(model_name="regnet_y_1_6gf", weights_name="RegNet_Y_1_6GF_Weights"),
+    TVModelMetadata(model_name="regnet_y_3_2gf", weights_name="RegNet_Y_3_2GF_Weights"),
+    TVModelMetadata(model_name="regnet_y_8gf", weights_name="RegNet_Y_8GF_Weights"),
+    TVModelMetadata(model_name="regnet_y_16gf", weights_name="RegNet_Y_16GF_Weights"),
+    TVModelMetadata(model_name="regnet_y_32gf", weights_name="RegNet_Y_32GF_Weights"),
+    TVModelMetadata(model_name="regnet_y_128gf", weights_name="RegNet_Y_128GF_Weights"),
+    TVModelMetadata(model_name="regnet_x_400mf", weights_name="RegNet_X_400MF_Weights"),
+    TVModelMetadata(model_name="regnet_x_800mf", weights_name="RegNet_X_800MF_Weights"),
+    TVModelMetadata(model_name="regnet_x_1_6gf", weights_name="RegNet_X_1_6GF_Weights"),
+    TVModelMetadata(model_name="regnet_x_3_2gf", weights_name="RegNet_X_3_2GF_Weights"),
+    TVModelMetadata(model_name="regnet_x_8gf", weights_name="RegNet_X_8GF_Weights"),
+    TVModelMetadata(model_name="regnet_x_16gf", weights_name="RegNet_X_16GF_Weights"),
+    TVModelMetadata(model_name="regnet_x_32gf", weights_name="RegNet_X_32GF_Weights"),
+    TVModelMetadata(model_name="swin_t", weights_name="Swin_T_Weights"),
+    TVModelMetadata(model_name="swin_s", weights_name="Swin_S_Weights"),
+    TVModelMetadata(model_name="swin_b", weights_name="Swin_B_Weights"),
+    TVModelMetadata(model_name="swin_v2_t", weights_name="Swin_V2_T_Weights"),
+    TVModelMetadata(
+        model_name="swin_v2_s", weights_name="Swin_V2_S_Weights", model_group="red"
+    ),
+    TVModelMetadata(model_name="swin_v2_b", weights_name="Swin_V2_B_Weights"),
 ]
 
 
 @pytest.mark.parametrize(
     "model_info",
-    model_info_list,
-    ids=[info[0] for info in model_info_list],
+    TORCHVISION_MODELS,
+    ids=lambda x: x.model_name,
 )
-@pytest.mark.parametrize("mode", ["train", "eval"])
+@pytest.mark.parametrize("mode", ["eval"])
 @pytest.mark.parametrize(
-    "op_by_op",
-    [OpByOpBackend.STABLEHLO, OpByOpBackend.TORCH, None],
-    ids=["op_by_op_stablehlo", "op_by_op_torch", "full"],
-)
-@pytest.mark.parametrize(
-    "data_parallel_mode", [False, True], ids=["single_device", "data_parallel"]
+    "execute_mode",
+    [CompileDepth.EXECUTE_OP_BY_OP, CompileDepth.EXECUTE],
+    ids=["op_by_op", "full"],
 )
 def test_torchvision_image_classification(
-    request, record_property, model_info, mode, op_by_op, data_parallel_mode
+    record_property,
+    model_info,
+    mode,
+    execute_mode,
 ):
-    if mode == "train":
-        pytest.skip()
-
     cc = CompilerConfig()
     cc.enable_consteval = True
     cc.consteval_parameters = True
-    if op_by_op:
-        if data_parallel_mode == "data_parallel":
-            pytest.skip("Op-by-op not supported in data parallel mode")
-        cc.compile_depth = CompileDepth.EXECUTE_OP_BY_OP
-        if op_by_op == OpByOpBackend.STABLEHLO:
-            cc.op_by_op_backend = OpByOpBackend.STABLEHLO
 
-    # TODO Enable checking (vit_h_14) - https://github.com/tenstorrent/tt-torch/issues/491
-    model_name = model_info[0]
-    assert_pcc = False if model_name in ["vit_h_14"] else True
-    assert_atol = False
+    # check if OpByOp
+    if execute_mode == CompileDepth.EXECUTE_OP_BY_OP:
+        cc.compile_depth = execute_mode
+    # applying overrides from model_metadata if EXECUTE
+    else:
+        cc.compile_depth = model_info.compile_depth
+    cc.op_by_op_backend = model_info.op_by_op_backend
 
-    model_group = "red" if model_name == "swin_v2_s" else "generality"
-
-    # Out of Memory: Not enough space to allocate 336691200 B DRAM buffer across 12 banks, where each bank needs to store 28057600 B
-    if model_name == "vit_h_14" and cc.compile_depth == CompileDepth.EXECUTE:
-        request.node.add_marker(
-            pytest.mark.xfail(
-                reason="Out of Memory: Not enough space to allocate in DRAM error - https://github.com/tenstorrent/tt-torch/issues/793",
-                strict=True,
-            )
-        )
+    skip_full_eval_test(
+        record_property,
+        cc,
+        model_info.model_name,
+        bringup_status="FAILED_RUNTIME",
+        reason="Out of Memory: Not enough space to allocate 336691200 B DRAM buffer across 12 banks, where each bank needs to store 28057600 B - https://github.com/tenstorrent/tt-torch/issues/793",
+        model_group=model_info.model_group,
+        model_name_filter=[
+            "vit_h_14",
+        ],
+    )
 
     tester = ThisTester(
-        model_info,
-        mode,
-        required_pcc=0.96,
-        assert_pcc=assert_pcc,
-        assert_atol=assert_atol,
+        model_name=model_info.model_name,
+        model_info=model_info,
+        mode=mode,
+        required_pcc=model_info.required_pcc,
+        assert_pcc=model_info.assert_pcc,
+        assert_atol=model_info.assert_atol,
         compiler_config=cc,
         record_property_handle=record_property,
-        model_group=model_group,
-        data_parallel_mode=data_parallel_mode,
+        model_group=model_info.model_group,
     )
     results = tester.test_model()
 
     def print_result(result):
         _, indices = torch.topk(result, 5)
-        print(f"Model: {model_name} | Top 5 predictions: {indices[0].tolist()}")
+        print(
+            f"Model: {model_info.model_name} | Top 5 predictions: {indices[0].tolist()}"
+        )
 
     if mode == "eval":
-        ModelTester.print_outputs(results, data_parallel_mode, print_result)
+        ModelTester.print_outputs(results, False, print_result)
 
     tester.finalize()
